Tests for deploying non-ugent VOs to scratch storage
